requirements.txt
fastapi
torch
torchvision
Pillow
uvicorn
pydantic
python-multipart